requirements.txt
absl-py==1.3.0
aiosignal==1.3.1
ale-py==0.7.5
attrs==22.1.0
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.2.0
cmake==3.26.4
contourpy==1.0.6
cycler==0.11.0
decorator==4.4.2
deepdiff==6.2.2
distlib==0.3.6
dm-env==1.5
dm-env-rpc==1.1.3
dm-tree==0.1.7
enum34==1.1.10
filelock==3.8.0
fonttools==4.38.0
frozenlist==1.3.3
googleapis-common-protos==1.57.0
grpcio==1.50.0
gym==0.25.1
gym-notices==0.0.8
gymnasium==0.27.0
gymnasium-notices==0.0.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
immutabledict==2.2.3
importlib-metadata==6.6.0
importlib-resources==5.10.1
jax-jumpy==0.2.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
kiwisolver==1.4.4
lit==16.0.6
lz4==4.0.2
MarkupSafe==2.1.3
matplotlib==3.6.2
mock==4.0.3
moviepy==1.0.3
mpmath==1.3.0
mpyq==0.2.5
msgpack==1.0.4
networkx==2.8.8
numpy==1.23.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.6.0.66
ordered-set==4.1.0
packaging==21.3
pandas==1.5.1
PettingZoo==1.20.1
Pillow==9.3.0
platformdirs==2.5.3
portpicker==1.5.2
proglog==0.1.10
protobuf==3.19.6
psutil==5.9.4
pygame==2.1.2
pyparsing==3.0.9
pyrsistent==0.19.2
PySC2==4.0.0
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
ray==2.1.0
requests==2.28.1
s2clientprotocol==5.0.10.89165.0
s2protocol==5.0.10.88500.0
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
Shimmy==0.1.0
six==1.16.0
sk-video==1.1.10
SuperSuit==3.5.0
sympy==1.12
tabulate==0.9.0
tensorboardX==2.5.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tinyscaler==1.2.5
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.64.1
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.6
websocket-client==1.4.2
zipp==3.15.0
